shared refactoring & iOS successful build
