+++ iosHyperskillApp/iosHyperskillApp/Sources/Modules/Step/StepViewModel.swift
@@ -10,15 +10,22 @@ final class StepViewModel: FeatureViewModel<StepFeatureState, StepFeatureMessage
 
     private let viewDataMapper: StepViewDataMapper
 
+    private let notificationService: NotificationsService
+    private let notificationsRegistrationService: NotificationsRegistrationService
+
     var stateKs: StepFeatureStateKs { .init(state) }
 
     init(
         stepRoute: StepRoute,
         viewDataMapper: StepViewDataMapper,
+        notificationService: NotificationsService,
+        notificationsRegistrationService: NotificationsRegistrationService,
         feature: Presentation_reduxFeature
     ) {
         self.stepRoute = stepRoute
         self.viewDataMapper = viewDataMapper
+        self.notificationService = notificationService
+        self.notificationsRegistrationService = notificationsRegistrationService
         super.init(feature: feature)
     }
 
@@ -60,6 +67,14 @@ final class StepViewModel: FeatureViewModel<StepFeatureState, StepFeatureMessage
         )
     }
 
+    func doDailyStepCompletedModalGoBackAction() {
+        onNewMessage(
+            StepFeatureMessageStepCompletionMessage(
+                message: StepCompletionFeatureMessageProblemOfDaySolvedModalGoBackClicked()
+            )
+        )
+    }
+
     private func updateStepQuiz() {
         mainScheduler.schedule { [weak self] in
             guard let strongSelf = self,
@@ -70,6 +85,33 @@ final class StepViewModel: FeatureViewModel<StepFeatureState, StepFeatureMessage
             strongSelf.stepQuizModuleInput?.update(
                 isPracticingLoading: stepFeatureStateData.stepCompletionState.isPracticingLoading
             )
+        }
+    }
+
+    // MARK: StepQuizUserPermissionRequest
+
+    func handleSendDailyStudyRemindersPermissionRequestResult(isGranted: Bool) {
+        let message = StepCompletionFeatureMessageRequestDailyStudyRemindersPermissionResult(
+            isGranted: isGranted
+        )
+
+        if isGranted {
+            Task(priority: .userInitiated) {
+                let isNotificationPermissionGranted =
+                await notificationsRegistrationService.requestAuthorizationIfNeeded()
+
+                await MainActor.run {
+                    onNewMessage(StepFeatureMessageStepCompletionMessage(message: message))
+
+                    if isNotificationPermissionGranted {
+                        notificationService.scheduleDailyStudyReminderLocalNotifications(
+                            analyticRoute: HyperskillAnalyticRoute.Learn.LearnStep(stepId: stepRoute.stepId)
+                        )
+                    }
+                }
+            }
+        } else {
+            onNewMessage(StepFeatureMessageStepCompletionMessage(message: message))
         }
     }
 
@@ -119,6 +161,24 @@ extension StepViewModel: TopicCompletedModalViewControllerDelegate {
         onNewMessage(
             StepFeatureMessageStepCompletionMessage(
                 message: StepCompletionFeatureMessageTopicCompletedModalHiddenEventMessage()
+            )
+        )
+    }
+}
+
+extension StepViewModel {
+    func logDailyStepCompletedModalShownEvent() {
+        onNewMessage(
+            StepFeatureMessageStepCompletionMessage(
+                message: StepCompletionFeatureMessageDailyStepCompletedModalShownEventMessage()
+            )
+        )
+    }
+
+    func logDailyStepCompletedModalHiddenEvent() {
+        onNewMessage(
+            StepFeatureMessageStepCompletionMessage(
+                message: StepCompletionFeatureMessageDailyStepCompletedModalHiddenEventMessage()
             )
         )
     }
