+++ iosHyperskillApp/iosHyperskillApp/Sources/Modules/StepQuiz/StepQuizViewModel.swift
@@ -17,10 +17,6 @@ final class StepQuizViewModel: FeatureViewModel<
     private var updateChildQuizSubscription: AnyCancellable?
 
     private let viewDataMapper: StepQuizViewDataMapper
-    private let userPermissionRequestTextMapper: StepQuizUserPermissionRequestTextMapper
-
-    private let notificationService: NotificationsService
-    private let notificationsRegistrationService: NotificationsRegistrationService
 
     var stateKs: StepQuizFeatureStateKs { .init(state) }
 
@@ -32,9 +28,6 @@ final class StepQuizViewModel: FeatureViewModel<
         moduleOutput: StepQuizOutputProtocol?,
         provideModuleInputCallback: @escaping (StepQuizInputProtocol?) -> Void,
         viewDataMapper: StepQuizViewDataMapper,
-        userPermissionRequestTextMapper: StepQuizUserPermissionRequestTextMapper,
-        notificationService: NotificationsService,
-        notificationsRegistrationService: NotificationsRegistrationService,
         feature: Presentation_reduxFeature
     ) {
         self.step = step
@@ -42,9 +35,6 @@ final class StepQuizViewModel: FeatureViewModel<
         self.moduleOutput = moduleOutput
         self.provideModuleInputCallback = provideModuleInputCallback
         self.viewDataMapper = viewDataMapper
-        self.userPermissionRequestTextMapper = userPermissionRequestTextMapper
-        self.notificationService = notificationService
-        self.notificationsRegistrationService = notificationsRegistrationService
         super.init(feature: feature)
     }
 
@@ -94,10 +84,6 @@ final class StepQuizViewModel: FeatureViewModel<
         moduleOutput?.stepQuizDidRequestContinue()
     }
 
-    func doGoBackAction() {
-        onNewMessage(StepQuizFeatureMessageProblemOfDaySolvedModalGoBackClicked())
-    }
-
     func makeViewData() -> StepQuizViewData {
         viewDataMapper.mapStepDataToViewData(step: step, state: state)
     }
@@ -119,62 +105,16 @@ final class StepQuizViewModel: FeatureViewModel<
         }
     }
 
-    // MARK: StepQuizUserPermissionRequest
-
-    func makeUserPermissionRequestTitle(_ userPermissionRequest: StepQuizUserPermissionRequest) -> String {
-        userPermissionRequestTextMapper.getTitle(request: userPermissionRequest)
-    }
-
-    func makeUserPermissionRequestMessage(_ userPermissionRequest: StepQuizUserPermissionRequest) -> String {
-        userPermissionRequestTextMapper.getMessage(request: userPermissionRequest)
-    }
-
     func handleResetCodePermissionRequestResult(isGranted: Bool) {
-        let message = StepQuizFeatureMessageRequestUserPermissionResult(
-            userPermissionRequest: StepQuizUserPermissionRequest.resetCode,
-            isGranted: isGranted
+        onNewMessage(
+            StepQuizFeatureMessageRequestResetCodeResult(isGranted: isGranted)
         )
-        onNewMessage(message)
-    }
-
-    func handleSendDailyStudyRemindersPermissionRequestResult(isGranted: Bool) {
-        let message = StepQuizFeatureMessageRequestUserPermissionResult(
-            userPermissionRequest: StepQuizUserPermissionRequest.sendDailyStudyReminders,
-            isGranted: isGranted
-        )
-
-        if isGranted {
-            Task(priority: .userInitiated) {
-                let isNotificationPermissionGranted =
-                  await notificationsRegistrationService.requestAuthorizationIfNeeded()
-
-                await MainActor.run {
-                    onNewMessage(message)
-
-                    if isNotificationPermissionGranted {
-                        notificationService.scheduleDailyStudyReminderLocalNotifications(
-                            analyticRoute: HyperskillAnalyticRoute.Learn.LearnStep(stepId: step.id)
-                        )
-                    }
-                }
-            }
-        } else {
-            onNewMessage(message)
-        }
     }
 
     // MARK: Analytic
 
     private func logClickedRetryEvent() {
         onNewMessage(StepQuizFeatureMessageClickedRetryEventMessage())
-    }
-
-    func logDailyStepCompletedModalShownEvent() {
-        onNewMessage(StepQuizFeatureMessageDailyStepCompletedModalShownEventMessage())
-    }
-
-    func logDailyStepCompletedModalHiddenEvent() {
-        onNewMessage(StepQuizFeatureMessageDailyStepCompletedModalHiddenEventMessage())
     }
 }
