+++ iosHyperskillApp/iosHyperskillApp/Sources/Modules/StepQuiz/Views/StepQuizView.swift
@@ -247,22 +247,8 @@ struct StepQuizView: View {
         switch StepQuizFeatureActionViewActionKs(viewAction) {
         case .showNetworkError:
             ProgressHUD.showError(status: Strings.General.connectionError)
-        case .requestUserPermission(let requestUserPermissionViewAction):
-            switch requestUserPermissionViewAction.userPermissionRequest {
-            case StepQuizUserPermissionRequest.resetCode:
-                presentResetCodePermissionAlert()
-            case StepQuizUserPermissionRequest.sendDailyStudyReminders:
-                presentSendDailyStudyRemindersPermissionAlert()
-            default:
-                break
-            }
-        case .showProblemOfDaySolvedModal(let showProblemOfDaySolvedModalViewAction):
-            presentDailyStepCompletedModal(earnedGemsText: showProblemOfDaySolvedModalViewAction.earnedGemsText)
-        case .navigateTo(let viewActionNavigateTo):
-            switch StepQuizFeatureActionViewActionNavigateToKs(viewActionNavigateTo) {
-            case .back:
-                stackRouter.popViewController()
-            }
+        case .requestResetCode:
+            presentResetCodePermissionAlert()
         }
     }
 }
@@ -272,8 +258,8 @@ struct StepQuizView: View {
 extension StepQuizView {
     private func presentResetCodePermissionAlert() {
         let alert = UIAlertController(
-            title: viewModel.makeUserPermissionRequestTitle(StepQuizUserPermissionRequest.resetCode),
-            message: viewModel.makeUserPermissionRequestMessage(StepQuizUserPermissionRequest.resetCode),
+            title: Strings.StepQuizCode.ResetCodePermissionAlert.title,
+            message: Strings.StepQuizCode.ResetCodePermissionAlert.message,
             preferredStyle: .alert
         )
         alert.addAction(
@@ -296,53 +282,5 @@ extension StepQuizView {
         )
 
         modalRouter.presentAlert(alert)
-    }
-
-    private func presentSendDailyStudyRemindersPermissionAlert() {
-        let alert = UIAlertController(
-            title: viewModel.makeUserPermissionRequestTitle(StepQuizUserPermissionRequest.sendDailyStudyReminders),
-            message: viewModel.makeUserPermissionRequestMessage(StepQuizUserPermissionRequest.sendDailyStudyReminders),
-            preferredStyle: .alert
-        )
-        alert.addAction(
-            UIAlertAction(
-                title: Strings.General.ok,
-                style: .default,
-                handler: { [weak viewModel] _ in
-                    viewModel?.handleSendDailyStudyRemindersPermissionRequestResult(isGranted: true)
-                }
-            )
-        )
-        alert.addAction(
-            UIAlertAction(
-                title: Strings.General.later,
-                style: .cancel,
-                handler: { [weak viewModel] _ in
-                    viewModel?.handleSendDailyStudyRemindersPermissionRequestResult(isGranted: false)
-                }
-            )
-        )
-
-        modalRouter.presentAlert(alert)
-    }
-}
-
-// MARK: - StepQuizView (Modals) -
-
-extension StepQuizView {
-    private func presentDailyStepCompletedModal(earnedGemsText: String) {
-        viewModel.logDailyStepCompletedModalShownEvent()
-
-        let panModal = ProblemOfDaySolvedModalViewController(
-            earnedGemsText: earnedGemsText,
-            onGoBackButtonTap: { [weak viewModel] in
-                viewModel?.doGoBackAction()
-            }
-        )
-        panModal.onDisappear = { [weak viewModel] in
-            viewModel?.logDailyStepCompletedModalHiddenEvent()
-        }
-
-        panModalPresenter.presentPanModal(panModal)
     }
 }
