+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_completion/injection/StepCompletionComponentImpl.kt
@@ -22,7 +22,8 @@ class StepCompletionComponentImpl(
             appGraph.analyticComponent.analyticInteractor,
             appGraph.commonComponent.resourceProvider,
             appGraph.sentryComponent.sentryInteractor,
-            appGraph.stepCompletionFlowDataComponent.topicCompletedFlow,
-            appGraph.buildNotificationComponent().notificationInteractor
+            appGraph.buildProfileDataComponent().profileInteractor,
+            appGraph.buildNotificationComponent().notificationInteractor,
+            appGraph.stepCompletionFlowDataComponent.topicCompletedFlow
         )
 }
