+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_completion/presentation/StepCompletionFeature.kt
@@ -55,10 +55,27 @@ interface StepCompletionFeature {
         }
 
         /**
+         * Show problem of day solve modal
+         */
+
+        data class ShowProblemOfDaySolvedModal(val earnedGemsText: String) : Message
+
+        object ProblemOfDaySolvedModalGoBackClicked : Message
+
+        /**
+         * Daily study reminders
+         */
+
+        object RequestDailyStudyRemindersPermission : Message
+        data class RequestDailyStudyRemindersPermissionResult(val isGranted: Boolean) : Message
+
+        /**
          * Analytic
          */
         object TopicCompletedModalShownEventMessage : Message
         object TopicCompletedModalHiddenEventMessage : Message
+        object DailyStepCompletedModalShownEventMessage : Message
+        object DailyStepCompletedModalHiddenEventMessage : Message
     }
 
     sealed interface Action {
@@ -68,8 +85,14 @@ interface StepCompletionFeature {
 
         data class CheckTopicCompletionStatus(val topicId: Long) : Action
 
+        data class RequestDailyStudyRemindersPermissionResult(val isGranted: Boolean) : Action
+
         sealed interface ViewAction : Action {
             data class ShowTopicCompletedModal(val modalText: String) : ViewAction
+
+            data class ShowProblemOfDaySolvedModal(val earnedGemsText: String) : ViewAction
+
+            object RequestDailyStudyRemindersPermission : ViewAction
 
             data class ShowStartPracticingError(val message: String) : ViewAction
