+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_completion/presentation/StepCompletionReducer.kt
@@ -10,6 +10,11 @@ import org.hyperskill.app.step_completion.presentation.StepCompletionFeature.Act
 import org.hyperskill.app.step_completion.presentation.StepCompletionFeature.ContinueButtonAction
 import org.hyperskill.app.step_completion.presentation.StepCompletionFeature.Message
 import org.hyperskill.app.step_completion.presentation.StepCompletionFeature.State
+import org.hyperskill.app.step_completion.domain.analytic.daily_notifications_notice.StepCompletionHiddenDailyNotificationsNoticeHyperskillAnalyticEvent
+import org.hyperskill.app.step_completion.domain.analytic.daily_notifications_notice.StepCompletionShownDailyNotificationsNoticeHyperskillAnalyticEvent
+import org.hyperskill.app.step_completion.domain.analytic.daily_step_completed_modal.StepCompletionDailyStepCompletedModalClickedGoBackHyperskillAnalyticEvent
+import org.hyperskill.app.step_completion.domain.analytic.daily_step_completed_modal.StepCompletionDailyStepCompletedModalHiddenHyperskillAnalyticEvent
+import org.hyperskill.app.step_completion.domain.analytic.daily_step_completed_modal.StepCompletionDailyStepCompletedModalShownHyperskillAnalyticEvent
 import ru.nobird.app.presentation.redux.reducer.StateReducer
 
 class StepCompletionReducer(private val stepRoute: StepRoute) : StateReducer<State, Message, Action> {
@@ -77,13 +82,32 @@ class StepCompletionReducer(private val stepRoute: StepRoute) : StateReducer<Sta
                         )
                     )
                 )
+            is Message.RequestDailyStudyRemindersPermission ->
+                state to setOf(
+                    Action.LogAnalyticEvent(
+                        StepCompletionShownDailyNotificationsNoticeHyperskillAnalyticEvent(stepRoute.analyticRoute)
+                    ),
+                    Action.ViewAction.RequestDailyStudyRemindersPermission
+                )
+            is Message.RequestDailyStudyRemindersPermissionResult -> {
+                val analyticEvent = StepCompletionHiddenDailyNotificationsNoticeHyperskillAnalyticEvent(
+                    route = stepRoute.analyticRoute,
+                    isAgreed = message.isGranted
+                )
+                state to setOf(
+                    Action.RequestDailyStudyRemindersPermissionResult(message.isGranted),
+                    Action.LogAnalyticEvent(analyticEvent)
+                )
+            }
             is Message.StepSolved ->
                 if (!state.isPracticingLoading &&
                     stepRoute is StepRoute.Learn &&
                     message.stepId == state.currentStep.id &&
                     state.currentStep.topic != null
                 ) {
-                    state.copy(isPracticingLoading = true) to setOf(Action.CheckTopicCompletionStatus(state.currentStep.topic))
+                    state.copy(isPracticingLoading = true) to setOf(
+                        Action.CheckTopicCompletionStatus(state.currentStep.topic)
+                    )
                 } else {
                     null
                 }
@@ -93,11 +117,34 @@ class StepCompletionReducer(private val stepRoute: StepRoute) : StateReducer<Sta
                 )
                 state to setOf(Action.LogAnalyticEvent(event))
             }
-            is Message.TopicCompletedModalHiddenEventMessage -> {
+            Message.TopicCompletedModalHiddenEventMessage -> {
                 val event = StepCompletionTopicCompletedModalHiddenHyperskillAnalyticEvent(
                     route = stepRoute.analyticRoute
                 )
                 state to setOf(Action.LogAnalyticEvent(event))
             }
+            is Message.DailyStepCompletedModalShownEventMessage ->
+                state to setOf(
+                    Action.LogAnalyticEvent(
+                        StepCompletionDailyStepCompletedModalShownHyperskillAnalyticEvent(stepRoute.analyticRoute)
+                    )
+                )
+            is Message.DailyStepCompletedModalHiddenEventMessage ->
+                state to setOf(
+                    Action.LogAnalyticEvent(
+                        StepCompletionDailyStepCompletedModalHiddenHyperskillAnalyticEvent(stepRoute.analyticRoute)
+                    )
+                )
+            is Message.ShowProblemOfDaySolvedModal ->
+                state to setOf(Action.ViewAction.ShowProblemOfDaySolvedModal(message.earnedGemsText))
+            is Message.ProblemOfDaySolvedModalGoBackClicked ->
+                state to setOf(
+                    Action.LogAnalyticEvent(
+                        StepCompletionDailyStepCompletedModalClickedGoBackHyperskillAnalyticEvent(
+                            stepRoute.analyticRoute
+                        )
+                    ),
+                    Action.ViewAction.NavigateTo.Back
+                )
         } ?: (state to emptySet())
 }
