--- shared/src/commonMain/kotlin/org/hyperskill/app/step_quiz/domain/model/permissions/StepQuizUserPermissionRequest.kt
@@ -1,6 +1,0 @@
-package org.hyperskill.app.step_quiz.domain.model.permissions
-
-enum class StepQuizUserPermissionRequest {
-    RESET_CODE,
-    SEND_DAILY_STUDY_REMINDERS
-}
