+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_quiz/injection/StepQuizComponent.kt
@@ -4,13 +4,11 @@ import org.hyperskill.app.step_quiz.domain.interactor.StepQuizInteractor
 import org.hyperskill.app.step_quiz.presentation.StepQuizFeature
 import org.hyperskill.app.step_quiz.view.mapper.StepQuizStatsTextMapper
 import org.hyperskill.app.step_quiz.view.mapper.StepQuizTitleMapper
-import org.hyperskill.app.step_quiz.view.mapper.StepQuizUserPermissionRequestTextMapper
 import ru.nobird.app.presentation.redux.feature.Feature
 
 interface StepQuizComponent {
     val stepQuizStatsTextMapper: StepQuizStatsTextMapper
     val stepQuizTitleMapper: StepQuizTitleMapper
-    val stepQuizUserPermissionRequestTextMapper: StepQuizUserPermissionRequestTextMapper
     val stepQuizInteractor: StepQuizInteractor
     val stepQuizFeature: Feature<StepQuizFeature.State, StepQuizFeature.Message, StepQuizFeature.Action>
 }
