+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_quiz/presentation/StepQuizFeature.kt
@@ -5,7 +5,6 @@ import org.hyperskill.app.profile.domain.model.Profile
 import org.hyperskill.app.step.domain.model.Step
 import org.hyperskill.app.step.domain.model.StepContext
 import org.hyperskill.app.step_quiz.domain.model.attempts.Attempt
-import org.hyperskill.app.step_quiz.domain.model.permissions.StepQuizUserPermissionRequest
 import org.hyperskill.app.step_quiz.domain.model.submissions.Reply
 import org.hyperskill.app.step_quiz.domain.model.submissions.Submission
 import org.hyperskill.app.step_quiz.domain.validation.ReplyValidationResult
@@ -69,29 +68,16 @@ interface StepQuizFeature {
         data class SyncReply(val reply: Reply) : Message
 
         /**
-         * Request user permission
+         * Reset code
          */
-        data class RequestUserPermission(val userPermissionRequest: StepQuizUserPermissionRequest) : Message
-        data class RequestUserPermissionResult(
-            val userPermissionRequest: StepQuizUserPermissionRequest,
-            val isGranted: Boolean
-        ) : Message
-
-        /**
-         * Show problem of day solve modal
-         */
-
-        data class ShowProblemOfDaySolvedModal(val earnedGemsText: String) : Message
-
-        object ProblemOfDaySolvedModalGoBackClicked : Message
+        object RequestResetCode : Message
+        data class RequestResetCodeResult(val isGranted: Boolean) : Message
 
         /**
          * Analytic
          */
         object ClickedCodeDetailsEventMessage : Message
         object ClickedRetryEventMessage : Message
-        object DailyStepCompletedModalShownEventMessage : Message
-        object DailyStepCompletedModalHiddenEventMessage : Message
     }
 
     sealed interface Action {
@@ -111,11 +97,6 @@ interface StepQuizFeature {
             val submission: Submission
         ) : Action
 
-        data class RequestUserPermissionResult(
-            val userPermissionRequest: StepQuizUserPermissionRequest,
-            val isGranted: Boolean
-        ) : Action
-
         /**
          * Analytic
          */
@@ -124,13 +105,7 @@ interface StepQuizFeature {
         sealed interface ViewAction : Action {
             object ShowNetworkError : ViewAction // error
 
-            data class RequestUserPermission(val userPermissionRequest: StepQuizUserPermissionRequest) : ViewAction
-
-            data class ShowProblemOfDaySolvedModal(val earnedGemsText: String) : ViewAction
-
-            sealed interface NavigateTo : ViewAction {
-                object Back : NavigateTo
-            }
+            object RequestResetCode : ViewAction
         }
     }
 }
